# lib/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamento v2",
    "version": "1.0.0",
    "description": "API REST completa para sistema de agendamentos com autenticação via API Key",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key no formato: sk_<apiClientId>_<random>"
      },
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token do Supabase (apenas para endpoints administrativos)"
      }
    },
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "integer"
          },
          "price": {
            "type": "number",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/services/{id}": {
      "get": {
        "tags": ["Agendamentos"],
        "summary": "Buscar Service por ID",
        "description": "Busca um service específico por ID na company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X GET http://localhost:3000/api/v1/services/123e4567-e89b-12d3-a456-426614174000 \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\"\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Service encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Service"}
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Consulta Médica",
                    "durationMinutes": 30,
                    "price": 150.0,
                    "created_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Service não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Service not found"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Agendamentos"],
        "summary": "Atualizar Service",
        "description": "Atualiza um service existente na company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X PUT http://localhost:3000/api/v1/services/123e4567-e89b-12d3-a456-426614174000 \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"name\": \"Consulta Médica Atualizada\",\n    \"durationMinutes\": 45,\n    \"price\": 200.0\n  }'\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Consulta Médica"
                  },
                  "durationMinutes": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 30
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true,
                    "example": 150.0
                  }
                }
              },
              "example": {
                "name": "Consulta Médica Atualizada",
                "durationMinutes": 45,
                "price": 200.0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Service"}
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Consulta Médica Atualizada",
                    "durationMinutes": 45,
                    "price": 200.0,
                    "created_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Validation error",
                  "errors": {
                    "durationMinutes": ["Number must be positive"],
                    "price": ["Number must be positive"]
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Service não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Service not found"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Agendamentos"],
        "summary": "Deletar Service",
        "description": "Deleta um service da company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X DELETE http://localhost:3000/api/v1/services/123e4567-e89b-12d3-a456-426614174000 \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\"\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Service deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {"type": "string", "example": "Service deleted successfully"}
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "message": "Service deleted successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Service não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Service not found"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}